FAQ_LLM_Telegram_Bot/requirements.txt
emoji==2.13.0
loguru==0.7.0
pyTelegramBotAPI==4.22.1
python-dotenv==1.0.1
google-genai==0.3.0
